In pinned mode, prevent user moving the panel
* tether hint can be moved vertically
