+++ macos/Onit/PanelStateManager/PanelStateManagerLogic.swift
@@ -10,6 +10,7 @@ protocol PanelStateManagerLogic {
     
     // MARK: - Properties
     
+    var isPanelMovable: Bool { get }
     var state: OnitPanelState { get set }
     var states: [OnitPanelState] { get set }
     var tetherButtonPanelState: OnitPanelState? { get set }
