+++ macos/Onit/PanelStateManager/Pinned/PanelStatePinnedManager.swift
@@ -31,6 +31,8 @@ class PanelStatePinnedManager: PanelStateBaseManager, ObservableObject {
     }
 
     // MARK: - PanelStateManagerLogic
+    
+    override var isPanelMovable: Bool { false }
 
     override func start() {
         stop()
