+++ macos/Onit/PanelStateManager/Untethered/PanelStateUntetheredManager.swift
@@ -30,15 +30,6 @@ class PanelStateUntetheredManager: PanelStateBaseManager, ObservableObject {
     }
 
     var tutorialWindow: NSWindow
-
-    private var shouldShowOnboarding: Bool {
-        let accessibilityPermissionGranted = AccessibilityPermissionManager.shared.accessibilityPermissionStatus == .granted
-        return !accessibilityPermissionGranted && Defaults[.showOnboarding]
-    }
-
-    // Dictionary that tracks whether the ExternalTetherButton should be visible for each screen.
-    // (For example, if the panel is closed then the tether button is visible.)
-    @Published var tetherButtonVisibility: [NSScreen: Bool] = [:]
     
     // MARK: - Initializer
     
@@ -107,7 +98,6 @@ class PanelStateUntetheredManager: PanelStateBaseManager, ObservableObject {
         
         lastScreenFrame = CGRect.zero
         statesByScreen = [:]
-        tetherButtonVisibility = [:]
     }
     
     override func getState(for windowHash: UInt) -> OnitPanelState? {
@@ -187,120 +177,6 @@ class PanelStateUntetheredManager: PanelStateBaseManager, ObservableObject {
         panelState.addDelegate(self)
         state = panelState
         return panelState
-    }
-
-
-    func debouncedShowTetherWindow(state: OnitPanelState, activeScreen: NSScreen) {
-        hideTetherWindow()
-        let pendingTetherWindow = (state, activeScreen)
-
-        tetherHintDetails.showTetherDebounceTimer = Timer.scheduledTimer(withTimeInterval: tetherHintDetails.showTetherDebounceDelay, repeats: false) { [weak self] _ in
-            guard let self = self else { return }
-            
-            DispatchQueue.main.async {
-                self.showTetherWindow(state: pendingTetherWindow.0, activeScreen: pendingTetherWindow.1)
-            }
-        }
-    }
-
-    func showTetherWindow(state: OnitPanelState, activeScreen: NSScreen) {
-         let tetherView = ExternalTetheredButton(
-            onClick: {
-                state.launchPanel()
-            },
-            onDrag: { [weak self] translation in
-                self?.tetheredWindowMoved(screen: activeScreen, y: translation)
-            }
-         ).environment(\.windowState, state)
-
-        let buttonView = NSHostingView(rootView: tetherView)
-        tetherHintDetails.tetherWindow.contentView = buttonView
-        tetherHintDetails.lastYComputed = nil
-        tetherButtonPanelState = state
-
-        if shouldShowOnboarding {
-            let tutorialView = TetherTutorialOverlay()
-            tutorialWindow.contentView = NSHostingView(rootView: tutorialView)
-            tutorialWindow.orderFrontRegardless()
-        }
-
-        updateTetherWindowPosition(for: activeScreen, lastYComputed: tetherHintDetails.lastYComputed)
-        tetherHintDetails.tetherWindow.orderFrontRegardless()
-    }
-    
-    private func updateTetherWindowPosition(for screen: NSScreen, lastYComputed: CGFloat? = nil) {
-        let activeScreenFrame = screen.visibleFrame
-        let positionX = activeScreenFrame.maxX - ExternalTetheredButton.containerWidth
-        var positionY: CGFloat
-        
-        if lastYComputed == nil {
-            positionY = activeScreenFrame.minY + (activeScreenFrame.height / 2) - (ExternalTetheredButton.containerHeight / 2)
-        } else {
-            positionY = computeHintYPosition(for: activeScreenFrame, offset: lastYComputed)
-        }
-        
-        let frame = NSRect(
-            x: positionX,
-            y: positionY,
-            width: ExternalTetheredButton.containerWidth,
-            height: ExternalTetheredButton.containerHeight
-        )
-        tetherHintDetails.tetherWindow.setFrame(frame, display: false)
-        
-        if (shouldShowOnboarding) {
-            let tutorialFrame = NSRect(
-                x: positionX - (TetherTutorialOverlay.width) + (ExternalTetheredButton.containerWidth / 2),
-                y: positionY + (ExternalTetheredButton.containerHeight / 2) - ((TetherTutorialOverlay.height * 1.5) / 2),
-                width: (TetherTutorialOverlay.width * 1.5),
-                height: (TetherTutorialOverlay.height * 1.5)
-            )
-            tutorialWindow.setFrame(tutorialFrame, display: false)
-        }
-    }
-    
-    private func computeHintYPosition(for screenVisibleFrame: CGRect, offset: CGFloat?) -> CGFloat {
-        let maxY = screenVisibleFrame.maxY - ExternalTetheredButton.containerHeight
-        let minY = screenVisibleFrame.minY
-
-        var lastYComputed = tetherHintDetails.lastYComputed ?? screenVisibleFrame.minY + (screenVisibleFrame.height / 2) - (ExternalTetheredButton.containerHeight / 2)
-
-        if let offset = offset {
-            lastYComputed -= offset
-        }
-
-        let finalOffset: CGFloat
-
-        if lastYComputed > maxY {
-            finalOffset = maxY
-        } else if lastYComputed < minY {
-            finalOffset = minY
-        } else {
-            finalOffset = lastYComputed
-        }
-
-        return finalOffset
-    }
-
-    private func tetheredWindowMoved(screen: NSScreen, y: CGFloat) {
-        let screenFrame = screen.visibleFrame
-        let lastYComputed = computeHintYPosition(for: screenFrame, offset: y)
-        
-        tetherHintDetails.lastYComputed = lastYComputed
-        
-        if let state = tetherButtonPanelState {
-            state.tetheredButtonYPosition = screenFrame.height -
-                (lastYComputed - screenFrame.minY) -
-                ExternalTetheredButton.containerHeight + (TetheredButton.height / 2)
-        }
-
-        let frame = NSRect(
-            x: tetherHintDetails.tetherWindow.frame.origin.x,
-            y: lastYComputed,
-            width: ExternalTetheredButton.containerWidth,
-            height: ExternalTetheredButton.containerHeight
-        )
-        
-        tetherHintDetails.tetherWindow.setFrame(frame, display: true)
     }
 }
